requirements.txt
asgiref==3.8.1
Django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
sqlparse==0.5.3
psycopg2-binary==2.9.9
dj-database-url==2.1.0
